requirements.txt
pyperclip==1.7.0
requests==2.21.0
numpy==1.15.4
Pillow==5.3.0
PyQt4==4.11.4
tensorflow==1.14.0
captcha==0.3
